Create project structure
Добавил структуру проекта
