+++ iOS10-HW14-Alexandr-Ivanov/App/AppDelegate.swift
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.swift
+//  iOS10-HW14-Alexandr-Ivanov
+//
+//  Created by Александр Иванов on 03.07.2023.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+}
