+++ iOS10-HW14-Alexandr-Ivanov/Resources/Assets.xcassets/Contents.json
