+++ iOS10-HW14-Alexandr-Ivanov/Sources/MainViewController.swift
@@ -1,0 +1,37 @@
+//
+//  MainViewController.swift
+//  iOS10-HW14-Alexandr-Ivanov
+//
+//  Created by Александр Иванов on 03.07.2023.
+//
+
+import UIKit
+import SnapKit
+
+class MainViewController: UIViewController {
+
+    // MARK: - Outlets
+
+    // MARK: - Lifecycle
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupView()
+        setupHeirarchy()
+        setupLayout()
+    }
+
+    // MARK: - Setup
+
+    private func setupView() {
+        view.backgroundColor = .white
+    }
+
+    private func setupHeirarchy() {
+
+    }
+
+    private func setupLayout() {
+
+    }
+}
