--- iOS10-HW14-Alexandr-Ivanov/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  iOS10-HW14-Alexandr-Ivanov
-//
-//  Created by Александр Иванов on 03.07.2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
